requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.27
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
psycopg[binary]==3.18.0
pyjwt==2.8.1
bcrypt==4.1.2
bleach==6.1.0
openai==1.13.3
redis==5.0.1
rq==1.15.1
pypdf==4.1.1
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0
pillow==10.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.25.2
